User class can be a STI subtype
Use active_support's `class_attribute` for the adapter, which allows
`rolify` to be called on the User base class, and roles to be used on a
subclass.

Previously there was a problem where the rolify code would call
`self.class.adapter` on the subclass which would fail. class_attribute
fixes this by allowing subclasses to the the base class's adapter value.