Add validation on Role.resource_type
With this will avoid add a permission to a model that don't call the
`resourcify`.

Now every model that call `resourcify` has his name saved and can be
found on the `Rolify.resource_types`.

Since the Role model is generated by the generator it is possible to
remove this validation.