Add eager loading to load_dynamic_methods.
Without this, every time a Rails process is started there is a single hit to the DB for each and every role for every single resource, e.g. if there are 5 roles for each of 100,000 resources, that's a half a million DB hits. At least in my stack, with 3 roles and a couple hundred resources, I'm seeing hundreds of DB hits on startup.