Update initializer.rb to include the remove_role_if_empty option and description
The default action is to remove roles if the last resource is removed. Important to disable this if the roles are associated with permissions (Pundit etc.) and cascades are enabled.

I'd suggest changing the default action, but in absence of that, a small change to the initializer to include the option and description to change it.