Enable short-circuiting in has_cached_role
Rather than iterate through the entire roles array and then report if
there are any values returned, we can instead use any? with a block
that will short-circuit when the first truthy value is returned.

This changes the call signature for RoleAdapter#find_cached* to return
boolean rather than an array